+++ CHANGELOG.md
@@ -4,6 +4,15 @@ To install or update CocoaPods see this [guide](http://docs.cocoapods.org/guides
 
 To install release candidates run `[sudo] gem install cocoapods --pre`
 
+## Master
+
+##### Enhancements
+
+* Show the name of the source for each hook that is run in verbose mode.  
+  [Samuel Giddins](https://github.com/segiddins)
+  [#2639](https://github.com/CocoaPods/CocoaPods/issues/2639)
+
+
 ## 0.35.0
 
 [CocoaPods](https://github.com/CocoaPods/CocoaPods/compare/0.34.4...0.35.0)